+++ components/common/Tabs.js
@@ -2,10 +2,14 @@
 import { useRouter } from "next/router";
 import { useState, useEffect } from "react";
 import { slugify } from "../../lib/slugify";
-import design from "../../styles/components/Window.module.sass"
+import design from "../../styles/components/Tabs.module.sass"
 
+const TabIndex = 0;
 export const Tabs = ({children, initialTab}) => {
   const [activeTab, setActiveTab] = useState(children[0].props.label);
+  console.log(TabIndex);
+  console.log(children[0]);
+  children[0].key = TabIndex + 1;
   const router = useRouter();
   const handleClick = (e, newActiveTab) => {
     e.preventDefault();
