+++ styles/components/Tabs.module.sass
